Bug fix- final card never received
When the deck is has 1 card left, it pops the last element from the deck into the players hand, but then because deck is empty, it then enters the next if statement (if the deck is empty), and so it then removes the card that was just given to the player. This effectively meant the deck had 49 cards, and was messing up my code for ages until I found it.